+++ lib/src/char_matcher/char_match.dart
@@ -1,0 +1,33 @@
+import 'package:meta/meta.dart';
+
+import '../../char_matcher.dart';
+
+@immutable
+class CharMatch extends Match {
+  @override
+  final int start;
+
+  @override
+  int get end => start + 1;
+
+  @override
+  final String input;
+
+  @override
+  final CharMatcher pattern;
+
+  CharMatch(this.start, this.input, this.pattern);
+
+  @override
+  String? group(int group) => this[group];
+
+  @override
+  String? operator [](int group) => group == 0 ? input[start] : null;
+
+  @override
+  List<String?> groups(List<int> groupIndices) =>
+      groupIndices.map(group).toList(growable: false);
+
+  @override
+  int get groupCount => 0;
+}