+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Optimize the implementation of `ExpiryCache`.
 * Fix `maxAge` of the `Stream.buffer` operator.
 * Add `Fraction.sign` and `Complex.sign`.
+* `CharMatcher` is now a `Pattern`.
 
 ## 2.8.0